% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{admixture}
\alias{admixture}
\title{admixture - estimate admixture components using supervised ADMIXTURE algorithm.}
\usage{
admixture(x, vr.ref)
}
\arguments{
\item{x}{myvcf of interest}

\item{vr.ref}{reference file for admixture estimation (uses subset of 1000 Genomes)}
}
\value{
List, name of metric (admixture), list of true false values if passed, ancestry coefficient values
}
\description{
admixture - estimate admixture components using supervised ADMIXTURE algorithm.
}
\keyword{internal}

