% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{callbyChrPlot}
\alias{callbyChrPlot}
\title{Dot plot of variant call counts (hom alt and het) by chromosome}
\usage{
callbyChrPlot(x, fn)
}
\arguments{
\item{x}{myvcf of interest}

\item{fn}{filename for plot title}
}
\value{
ggplot of variant calls by chromosome
}
\description{
Dot plot of variant call counts (hom alt and het) by chromosome
}
\keyword{internal}
