% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{computeTiTv}
\alias{computeTiTv}
\title{Private function to calc transition tranversion (titv) ratio}
\usage{
computeTiTv(x)
}
\arguments{
\item{x}{Table of SNPs}
}
\value{
Transition Transversion Ratio
}
\description{
Private function to calc transition tranversion (titv) ratio
}
\keyword{internal}
