% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{genotypeQualityPlot}
\alias{genotypeQualityPlot}
\title{Histogram of genotype qualities}
\usage{
genotypeQualityPlot(x, fn)
}
\arguments{
\item{x}{myvcf of interest}

\item{fn}{filename for plot title}
}
\value{
genotype quality histogram as ggplot
}
\description{
Histogram of genotype qualities
}
\keyword{internal}
