% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{getCoefs}
\alias{getCoefs}
\title{Private function to calc coefficients for admixture}
\usage{
getCoefs(seed, dat)
}
\arguments{
\item{seed}{seed to start constroptim}

\item{dat}{current data frame}
}
\value{
Ancestry coefficients
}
\description{
Private function to calc coefficients for admixture
}
\keyword{internal}

