% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{getDescriptors}
\alias{getDescriptors}
\title{Return vector of long descriptions for variable names}
\usage{
getDescriptors(x)
}
\arguments{
\item{x}{variable names to translate}
}
\value{
description of variable names
}
\description{
Return vector of long descriptions for variable names
}
\keyword{internal}
