% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{getName}
\alias{getName}
\title{Getter for VCFQAReport class to return filename slot}
\usage{
getName(Object)
}
\arguments{
\item{Object}{Object of class VCFQAReport}
}
\value{
Name of file
}
\description{
Getter for VCFQAReport class to return filename slot
}
\examples{
vcffn <- system.file("ext-data", "chr22.GRCh38.vcf.gz", package="genotypeeval")
mydir <- paste(dirname(vcffn), "/", sep="")
myfile <-basename(vcffn)
svp <- ScanVcfParam(which=GRanges("22", IRanges(0,200e5)), geno="GT")
vcfparams <- VCFQAParam(count.limits=c(3014580000, Inf), readdepth.target = 30)
vcf <- ReadVCFData(mydir, myfile, "GRCh38")
ev <- VCFEvaluate(vcf, vcfparams)
getName(ev)
}
