% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{getPlots}
\alias{getPlots}
\title{Getter for VCFQAReport class to return plots slot.}
\usage{
getPlots(Object)
}
\arguments{
\item{Object}{Object of Class VCFQAReport}
}
\value{
List of named ggplots
}
\description{
Getter for VCFQAReport class to return plots slot.
}
\examples{
vcffn <- system.file("ext-data", "chr22.GRCh38.vcf.gz", package="genotypeeval")
mydir <- paste(dirname(vcffn), "/", sep="")
myfile <-basename(vcffn)
svp <- ScanVcfParam(which=GRanges("22", IRanges(0,200e5)), geno="GT")
vcfparams <- VCFQAParam(count.limits=c(3014580000, Inf), readdepth.target = 30)
vcf <- ReadVCFData(mydir, myfile, "GRCh38")
ev <- VCFEvaluate(vcf, vcfparams)
getPlots(ev)
}

