% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{getResults}
\alias{getResults}
\title{Getter for VCFQAReport class to return results.  Return a list showing values that the sample was evaluated on.}
\usage{
getResults(Object)
}
\arguments{
\item{Object}{an object of type VCFQAReport}
}
\value{
numeric vector of results
}
\description{
Getter for VCFQAReport class to return results.  Return a list showing values that the sample was evaluated on.
}
\examples{
vcffn <- system.file("ext-data", "chr22.GRCh38.vcf.gz", package="genotypeeval")
mydir <- paste(dirname(vcffn), "/", sep="")
myfile <-basename(vcffn)
svp <- ScanVcfParam(which=GRanges("22", IRanges(0,200e5)), geno="GT")
vcfparams <- VCFQAParam(count.limits=c(3014580000, Inf), readdepth.target = 30)
vcf <- ReadVCFData(mydir, myfile, "GRCh38")
ev <- VCFEvaluate(vcf, vcfparams)
getResults(ev)
}
