% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFData.R
\name{getVR}
\alias{getVR}
\title{getVr is a Getter. Returns vr slot.}
\usage{
getVR(x)
}
\arguments{
\item{x}{VCFData object}
}
\value{
VRanges
}
\description{
getVr is a Getter. Returns vr slot.
}
\examples{
vcffn <- system.file("ext-data", "chr22.GRCh38.vcf.gz", package="genotypeeval")
mydir <- paste(dirname(vcffn), "/", sep="")
myfile <-basename(vcffn)
svp <- ScanVcfParam(which=GRanges("22", IRanges(0,1e5)), geno="GT")
vcf <- ReadVCFData(mydir, myfile, "GRCh38")
getVR(vcf)
}

