% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{goldCompare}
\alias{goldCompare}
\title{Comparator to gold standard}
\usage{
goldCompare(x, gr, coding, limits.c, limits.uc, percent.c, name.of.gold)
}
\arguments{
\item{x}{myvcf of interest}

\item{gr}{gold GRanges}

\item{coding}{Coding GRanges}

\item{limits.c}{all confirmed limits}

\item{limits.uc}{all unconfirmed limits}

\item{percent.c}{percent confirmed limit}

\item{name.of.gold}{user readable name}
}
\value{
list with name, logical vector (passed or not), numeric vector of transition transversion in coding and non coding regions in confirmed and non confirmed
}
\description{
Comparator to gold standard
}
\keyword{internal}

