% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldData.R
\docType{class}
\name{GoldData-class}
\alias{GoldData-class}
\title{Declare class Gold to store information from Gold" (1000 Genomes for example) along with the GoldDataParam}
\arguments{
\item{genome}{Genome build, GRCh37 or GRCh38}

\item{track}{Where the gold data is stored}

\item{goldparams}{The Param file with the limits to be applied}

\item{track.rare}{Stores the Gold data with MAF < 0.01 if MAF exists}
}
\value{
Object of class GoldData
}
\description{
Declare class Gold to store information from Gold" (1000 Genomes for example) along with the GoldDataParam
}

