% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldData.R
\name{GoldDataFromGRanges}
\alias{GoldDataFromGRanges}
\title{User Constructor for class.  Used to associate the gold params object with the gold granges and to check if MAF is present.}
\usage{
GoldDataFromGRanges(genome, gold.granges, goldparams)
}
\arguments{
\item{genome}{Genome build, GRCh37 or GRCh38}

\item{gold.granges}{Gold file as GRanges}

\item{goldparams}{GoldDataParam object setting thresholds for evaluation}
}
\value{
Object of class GoldData
}
\description{
User Constructor for class.  Used to associate the gold params object with the gold granges and to check if MAF is present.
}
\examples{
gparam <-  GoldDataParam(percent.confirmed=0.792, percent.het.rare = 0.93)
gr <- GRanges(seqnames="22", IRanges(1e7,5e7))
gold <- GoldDataFromGRanges("GRCh38", gr, gparam)
}
