% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldDataParam.R
\docType{class}
\name{GoldDataParam-class}
\alias{GoldDataParam-class}
\title{Declare class GoldDataParam which will store thresholds to apply to VCFEvaluate object.  This is intended for use in batch mode when a large number of vcf files needs to be screened and individual vcf files that fail flagged.
All limits follow the format lower limit than upper limit}
\arguments{
\item{titv.confirmed.limits}{lower limit coding, upper limit coding, lower limit noncoding, upper limit noncoding, Transition transversion ratios for confirmed snps}

\item{titv.unconfirmed.limits}{lower limit coding, upper limit coding, lower limit noncoding, upper limit noncoding, Transition transversion ratios for unconfirmed snps}

\item{percent.confirmed.limits}{lower limit, upper limit, percent confirmed in Gold comparator}

\item{percent.het.rare.limits}{lower limit, upper limit, (Percent Het in Rare, MAF < 0.01 in Gold) / Total number of Heterozygotes}
}
\value{
Object of type GoldDataParam
}
\description{
Declare class GoldDataParam which will store thresholds to apply to VCFEvaluate object.  This is intended for use in batch mode when a large number of vcf files needs to be screened and individual vcf files that fail flagged.
All limits follow the format lower limit than upper limit
}

