% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldDataParam.R
\name{GoldDataParam}
\alias{GoldDataParam}
\title{User Constructor for class}
\usage{
GoldDataParam(titv.coding.confirmed.l = 0, titv.coding.confirmed.u = 5,
  titv.noncoding.confirmed.l = 0, titv.noncoding.confirmed.u = 5,
  titv.coding.unconfirmed.l = 0, titv.coding.unconfirmed.u = 5,
  titv.noncoding.unconfirmed.l = 0, titv.noncoding.unconfirmed.u = 5,
  percent.confirmed.limits = 0, percent.het.rare.limits = 0)
}
\arguments{
\item{titv.coding.confirmed.l}{Lower limit of transition transversion ratio in coding confirmed}

\item{titv.coding.confirmed.u}{upper limit of transition transverion ratio coding confirmed}

\item{titv.noncoding.confirmed.l}{Lower limit of transition transversion ratio in noncoding confirmed}

\item{titv.noncoding.confirmed.u}{upper limit of transition transverion ratio noncoding confirmed}

\item{titv.coding.unconfirmed.l}{Lower limit of transition transversion ratio in coding unconfirmed}

\item{titv.coding.unconfirmed.u}{upper limit of transition transverion ratio coding unconfirmed}

\item{titv.noncoding.unconfirmed.l}{Lower limit of transition transversion ratio in noncoding unconfirmed}

\item{titv.noncoding.unconfirmed.u}{upper limit of transition transverion ratio noncoding unconfirmed}

\item{percent.confirmed.limits}{lower limit, upper limit, percent confirmed in Gold comparator}

\item{percent.het.rare.limits}{lower limit, upper limit, (Percent Het in Rare, MAF < 0.01 in Gold) / Total number of Heterozygotes}
}
\value{
Object of type GoldDataParam
}
\description{
User Constructor for class
}
\examples{
gparam <- GoldDataParam(percent.confirmed=0.792, percent.het.rare = 0.93)
}

