% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{hetGap}
\alias{hetGap}
\title{Gap between HETs by Chromosome}
\usage{
hetGap(x, myLimits)
}
\arguments{
\item{x}{myvcf of interest}

\item{myLimits}{is largest allowable gap by chromosome}
}
\value{
list with name, logical vector (passed or not), numeric vector maximum gap between two het calls by chromosome
}
\description{
Gap between HETs by Chromosome
}
\keyword{internal}

