% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{hetsMasked}
\alias{hetsMasked}
\title{Number hets in masked GRanges}
\usage{
hetsMasked(x, masked.limits, non.masked.limits, masked)
}
\arguments{
\item{x}{myvcf of interest}

\item{masked.limits}{Limits for self chain region (lower and upper)}

\item{non.masked.limits}{Limits for non self chain region (lower and upper)}

\item{masked}{GRanges of difficult regions (repeat masked, paralogs, self chain, etc)}
}
\value{
list with name, logical vector (passed or not), numeric vector number hets in masked region and not in masked region
}
\description{
Number hets in masked GRanges
}
\keyword{internal}

