% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{homrefPlot}
\alias{homrefPlot}
\title{Dot plot of variant call counts (hom ref) by chromosome}
\usage{
homrefPlot(x, fn)
}
\arguments{
\item{x}{myvcf of interest}

\item{fn}{filename for plot title}
}
\value{
ggplot of homozygous reference counts
}
\description{
Dot plot of variant call counts (hom ref) by chromosome
}
\keyword{internal}

