% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{meanGQ}
\alias{meanGQ}
\title{Mean Genotype Quality (GQ)}
\usage{
meanGQ(x, myLimit)
}
\arguments{
\item{x}{myvcf of interest}

\item{myLimit}{From params, the bounds for Genotype Quality}
}
\value{
list with name, logical (passed or not), numeric (mean genotype quality)
}
\description{
Mean Genotype Quality (GQ)
}
\keyword{internal}

