% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{myf}
\alias{myf}
\title{Private function to calc likelihood for admixture}
\usage{
myf(data, par)
}
\arguments{
\item{data}{current data frame}

\item{par}{parameters to estimate}
}
\value{
numeric, Estimates of log likelihood of ancestry coefficients
}
\description{
Private function to calc likelihood for admixture
}
\keyword{internal}

