% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{numberCalls}
\alias{numberCalls}
\title{Total Calls.  This is the total number of calls in the file (including MULTIs so hom ref, hom var and hom alt might not add up).  These methods are private.  Users are not expected to provide the number of hom ref and number of duplicate calls - these functions are generally called through the function VCFEvaluate}
\usage{
numberCalls(x, n.hom, n.dup, ll, ul)
}
\arguments{
\item{x}{myvcf of interest}

\item{n.hom}{Number of homozygous reference.  Pass so don't need to recalculate}

\item{n.dup}{number of multi calls to add onto the total calls}

\item{ll}{lower limit to pass for hom ref}

\item{ul}{upper limit to pass for hom ref}
}
\value{
list with name, logical (passed or not), numeric (total number of calls)
}
\description{
Total Calls.  This is the total number of calls in the file (including MULTIs so hom ref, hom var and hom alt might not add up).  These methods are private.  Users are not expected to provide the number of hom ref and number of duplicate calls - these functions are generally called through the function VCFEvaluate
}
\keyword{internal}
