% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{numberOfHets}
\alias{numberOfHets}
\title{Count Number of Hets}
\usage{
numberOfHets(x, ll, ul)
}
\arguments{
\item{x}{myvcf of interest}

\item{ll}{lower limit to pass for hets}

\item{ul}{upper limit to pass for hets}
}
\value{
list with name, logical (passed or not), numeric (number of hets)
}
\description{
Count Number of Hets
}
\keyword{internal}

