% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{numberOfHomRefs}
\alias{numberOfHomRefs}
\title{Count Number of Hom Ref}
\usage{
numberOfHomRefs(x, ll, ul)
}
\arguments{
\item{x}{myvcf of interest}

\item{ll}{lower limit to pass for hom ref}

\item{ul}{upper limit to pass for hom ref}
}
\value{
list with name, logical (passed or not), numeric (number of hom ref)
}
\description{
Count Number of Hom Ref
}
\keyword{internal}

