% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{numberOfHomVars}
\alias{numberOfHomVars}
\title{Count Number of Hom Vars}
\usage{
numberOfHomVars(x, ll, ul)
}
\arguments{
\item{x}{myvcf of interest}

\item{ll}{lower limit to pass for Hom Vars}

\item{ul}{upper limit to pass for Hom Vars}
}
\value{
list with name, logical (passed or not), numeric (number of hom vars)
}
\description{
Count Number of Hom Vars
}
\keyword{internal}

