% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{percentHets}
\alias{percentHets}
\title{Percent of Hets as Total number of variants}
\usage{
percentHets(x, n.hom, ll, ul)
}
\arguments{
\item{x}{myvcf of interest}

\item{n.hom}{Pass the number of homozygous reference so don't need to recalculate}

\item{ll}{lower limit to pass for percent}

\item{ul}{upper limit to pass for percent}
}
\value{
list with name, logical (passed or not), numeric (percent hets)
}
\description{
Percent of Hets as Total number of variants
}
\keyword{internal}
