% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{rareCompare}
\alias{rareCompare}
\title{Comparator to rare variants.  Rare is defined as 0.01 percent or less}
\usage{
rareCompare(x, gr, percent.h)
}
\arguments{
\item{x}{myvcf of interest}

\item{gr}{GRanges of gold comparator}

\item{percent.h}{what percent higher hets than hom_alt?}
}
\value{
list with name, logical (passed or not), numeric (proportion hets in rare)
}
\description{
Comparator to rare variants.  Rare is defined as 0.01 percent or less
}
\keyword{internal}

