% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{readDepth}
\alias{readDepth}
\title{Median read depth}
\usage{
readDepth(x, coverage)
}
\arguments{
\item{x}{myvcf of interest}

\item{coverage}{median coverage cutoff}
}
\value{
list with name, logical (passed or not), numeric median read depth
}
\description{
Median read depth
}
\keyword{internal}

