% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{readdepthPlot}
\alias{readdepthPlot}
\title{Histogram of read depth by GT}
\usage{
readdepthPlot(x, fn)
}
\arguments{
\item{x}{myvcf of interest}

\item{fn}{The filename (for title of the plot)}
}
\value{
ggplot of histogram of read depth
}
\description{
Histogram of read depth by GT
}
\keyword{internal}

