% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldData.R
\name{ReadGoldData}
\alias{ReadGoldData}
\title{User Constructor for class}
\usage{
ReadGoldData(genome, vcffilename, goldparams)
}
\arguments{
\item{genome}{Genome build, GRCh37 or GRCh38}

\item{vcffilename}{path and filename of vcf file}

\item{goldparams}{GoldDataParam object setting thresholds for evaluation}
}
\value{
Object of class GoldData
}
\description{
User Constructor for class
}
\examples{
gparam <-  GoldDataParam(percent.confirmed=0.792, percent.het.rare = 0.93)
g1000fn <- system.file("ext-data", "example_gold_file.vcf", package="genotypeeval")
g1000 <- ReadGoldData("GRCh38", g1000fn, gparam)
}

