% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFData.R
\name{ReadVCFData}
\alias{ReadVCFData}
\title{User Constructor for class.  Calls VCFData constructor:
ReadVCFData is a wrapper for readVcfAsVRanges.  It removes indels, GL chromosomes, and MULTI calls.
It scans the header of the vcf file and adds in the following fields for analysis if present:  AD, GT, DP, GQ.
Looks for the "END" tag in the header and reads in file as gVCF if necessary.}
\usage{
ReadVCFData(mydir, myfile, genome)
}
\arguments{
\item{mydir}{Directory of vcf file}

\item{myfile}{Filename of vcf file}

\item{genome}{GRCh37 or GRCh38}
}
\value{
Object of class VCFData
}
\description{
User Constructor for class.  Calls VCFData constructor:
ReadVCFData is a wrapper for readVcfAsVRanges.  It removes indels, GL chromosomes, and MULTI calls.
It scans the header of the vcf file and adds in the following fields for analysis if present:  AD, GT, DP, GQ.
Looks for the "END" tag in the header and reads in file as gVCF if necessary.
}
\examples{
vcffn <- system.file("ext-data", "chr22.GRCh38.vcf.gz", package="genotypeeval")
mydir <- paste(dirname(vcffn), "/", sep="")
myfile <-basename(vcffn)
vcf <- ReadVCFData(mydir, myfile, "GRCh38")
}

