% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GoldData.R
\name{readVcfGold}
\alias{readVcfGold}
\title{Private method for class.  Read in Gold file - will read in the AF if it is detected in header}
\usage{
readVcfGold(vcfFile, rare, genome)
}
\arguments{
\item{vcfFile}{Gold file to read in}

\item{rare}{Whether AF is in the header}

\item{genome}{Build GRCh37 or GRCh38}
}
\value{
GRanges, read in from gold file
}
\description{
Private method for class.  Read in Gold file - will read in the AF if it is detected in header
}
\keyword{internal}

