% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationSummary.R
\name{reformatData}
\alias{reformatData}
\title{Take in the results from the population data and re-format it}
\usage{
reformatData(results)
}
\arguments{
\item{results}{The list of results from running the package using BatchJobs}
}
\value{
list, data frame of logical (passed or not), data frame of numeric (all results)
}
\description{
Take in the results from the population data and re-format it
}
\keyword{private}

