% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{titv}
\alias{titv}
\title{Transition transversion ratio in coding and non-coding}
\usage{
titv(x, coding, ll.nc, ul.nc, ll.c, ul.c)
}
\arguments{
\item{x}{myvcf of interest}

\item{coding}{GRanges of coding regions}

\item{ll.nc}{lower limit to pass for transition transversion in noncoding}

\item{ul.nc}{upper limit to pass for transition transversion in noncoding}

\item{ll.c}{lower limit to pass for transition transversion in coding}

\item{ul.c}{upper limit to pass for transition transversion in coding}
}
\value{
list with name, logical vector (passed or not), numeric vector of transition transversion ratio in coding and non coding regions
}
\description{
Transition transversion ratio in coding and non-coding
}
\keyword{internal}

