% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFData.R
\docType{class}
\name{VCFData-class}
\alias{VCFData-class}
\title{Declare class
Reads in VCF using readVCFAsVRanges}
\arguments{
\item{mydir}{Directory of vcf file}

\item{myfile}{Filename of vcf file}

\item{vr.homref}{All SNPs from VCF with INDELs, MULTIs (seperately removed for variant and non variant), weird chromosomes removed}

\item{genoString}{A character vector of all genotype fields present (looks for AD, GQ, GT, DP)}

\item{infoString}{A character vector looking for "END" tag indicating file is a gVCF}

\item{genome}{Declare if the genome is GRCh37 or GRCh38}

\item{n.dup}{Counts the number of MULTIs removed}

\item{chunked}{Whether data was read in using ReadVCFDataChunk which means hom refs not in the admixture file were dropped}
}
\value{
Object of class VCFData
}
\description{
Declare class
Reads in VCF using readVCFAsVRanges
}

