% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{VCFEvaluate}
\alias{VCFEvaluate}
\title{Constructor for class.  Calls constructor for class.  Using the GENO fields present in the vcf header will evaluate the vcf file using metrics and generate plots.  Each metric will be tested against the params specified in the params class.  For example, if Read Depth is in the GENO header will calculate median read depth, percent in target (50 percent to 200 percent of the target specified in the params file) and generate a histogram of Read Depth.}
\usage{
VCFEvaluate(myvcf, vcfparams, gold.ref = NA, cds.ref = NA,
  masked.ref = NA, admixture.ref = NA)
}
\arguments{
\item{myvcf}{Vcf file to evaluate}

\item{vcfparams}{object of VCFQAParam class.  Sets thresholds to evaluate the VCF File against.}

\item{gold.ref}{Object of class Gold that contains the 1000 Genomes reference}

\item{cds.ref}{Coding Region as GRanges}

\item{masked.ref}{optional regions as GRanges to mask eg repeats, self chain, paralogs, etc.}

\item{admixture.ref}{VRanges with MAF for superpopulations (EAS, AFR, EUR)}
}
\value{
Object of VCFQAReport.
}
\description{
Constructor for class.  Calls constructor for class.  Using the GENO fields present in the vcf header will evaluate the vcf file using metrics and generate plots.  Each metric will be tested against the params specified in the params class.  For example, if Read Depth is in the GENO header will calculate median read depth, percent in target (50 percent to 200 percent of the target specified in the params file) and generate a histogram of Read Depth.
}
\examples{
vcffn <- system.file("ext-data", "chr22.GRCh38.vcf.gz", package="genotypeeval")
mydir <- paste(dirname(vcffn), "/", sep="")
myfile <-basename(vcffn)
svp <- ScanVcfParam(which=GRanges("22", IRanges(0,200e5)), geno="GT")
vcfparams <- VCFQAParam(count.limits=c(3014580000, Inf), readdepth.target = 30)
vcf <- ReadVCFData(mydir, myfile, "GRCh38")
ev <- VCFEvaluate(vcf, vcfparams)
}
