% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAParam.R
\docType{class}
\name{VCFQAParam-class}
\alias{VCFQAParam-class}
\title{Declare class VCFQAParam which will store thresholds to apply to VCFEvaluate object.  This is intended for use in batch mode when a large number of vcf files needs to be screened and individual vcf files that fail flagged.
All limits follow the format lower limit than upper limit}
\arguments{
\item{homref.limits}{lower limit, upper limit, number of homozygous reference}

\item{het.limits}{lower limit, upper limit, number of heterozygous calls}

\item{homvar.limits}{lower limit, upper limit, number of homozygous alternative}

\item{count.limits}{lower limit, upper limit, total number of counts}

\item{percenthets.limits}{lower limit, upper limit, Number of Heterozgyous / (Total Number of Counts) or percent het}

\item{titv.noncoding.limits}{lower limit, upper limit, Transition transversion ratio in noncoding regions}

\item{titv.coding.limits}{lower limit, upper limit, Transition transversion ratio in coding regions}

\item{readdepth.target}{The sequencing depth target (eg 30x)}

\item{readdepth.limits}{lower limit, upper limit, Mean read depth}

\item{readdepth.percent.limits}{lower limit, upper limit, Percent read depth in target (50 percent to 200 percent of target read depth)}

\item{gq.limit}{lower limit, Mean genotype quality (does not make sense to have an upper limit)}

\item{masked.limits}{lower limit, upper limit, (Number of heterozygous in masked regions)/(Total number of heterozygotes)}

\item{non.masked.limits}{lower limit, upper limit, (Number of heterozygous in non-self chained regions)/(Total number of heterozygotes)}

\item{het.gap.limits}{lower limit, upper limit, Largest gap within chromosome between two heterozygous calls}

\item{gq.filter}{filter for the VCF file on genotype quality (eg only GQ > 90)}

\item{dp.filter}{filter for the VCF file on read depth (eg only DP > 0)}
}
\value{
VCFQAParam object
}
\description{
Declare class VCFQAParam which will store thresholds to apply to VCFEvaluate object.  This is intended for use in batch mode when a large number of vcf files needs to be screened and individual vcf files that fail flagged.
All limits follow the format lower limit than upper limit
}
