% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAParam.R
\name{VCFQAParam}
\alias{VCFQAParam}
\title{User Constructor for class.  Call limits are set as default to pass.}
\usage{
VCFQAParam(homref.limits = c(-Inf, Inf), het.limits = c(-Inf, Inf),
  homvar.limits = c(-Inf, Inf), percenthets.limits = c(-Inf, Inf),
  titv.noncoding.limits = c(-Inf, Inf), titv.coding.limits = c(-Inf, Inf),
  readdepth.target = -1, readdepth.limits = c(-Inf, Inf),
  readdepth.percent.limits = 0, gq.limit = 0, masked.limits = c(-Inf,
  Inf), non.masked.limits = c(-Inf, Inf), het.gap.limits = rep(Inf, 24),
  count.limits = c(-Inf, Inf), gq.filter = 0, dp.filter = -1)
}
\arguments{
\item{homref.limits}{lower limit, upper limit, number of homozygous reference}

\item{het.limits}{lower limit, upper limit, number of heterozygous calls}

\item{homvar.limits}{lower limit, upper limit, number of homozygous alternative}

\item{percenthets.limits}{lower limit, upper limit, Number of Heterozgyous / (Total Number of Counts) or percent het}

\item{titv.noncoding.limits}{lower limit, upper limit, Transition transversion ratio in noncoding regions}

\item{titv.coding.limits}{lower limit, upper limit, Transition transversion ratio in coding regions}

\item{readdepth.target}{The sequencing depth target (eg 30x)}

\item{readdepth.limits}{lower limit, upper limit, Mean read depth}

\item{readdepth.percent.limits}{lower limit, upper limit, Percent read depth in target (50 percent to 200 percent of target read depth)}

\item{gq.limit}{lower limit, Mean genotype quality (does not make sense to have an upper limit)}

\item{masked.limits}{lower limit, upper limit, (Number of heterozygous in self chained regions)/(Total number of heterozygotes)}

\item{non.masked.limits}{lower limit, upper limit, (Number of heterozygous in non-self chained regions)/(Total number of heterozygotes)}

\item{het.gap.limits}{lower limit, upper limit, Largest gap within chromosome between two heterozygous calls}

\item{count.limits}{lower limit, upper limit, total number of counts}

\item{gq.filter}{filter for the VCF file on genotype quality (eg only GQ > 90)}

\item{dp.filter}{filter for the VCF file on read depth (eg only DP > 0)}
}
\value{
Object of class VCFQAParam
}
\description{
User Constructor for class.  Call limits are set as default to pass.
}
\examples{
vcfparams <- VCFQAParam(count.limits=c(3014580000, Inf), readdepth.target = 30)
}
