% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\docType{class}
\name{VCFQAReport-class}
\alias{VCFQAReport-class}
\title{Declare class VCFQAReport which will evaluate a VCF stored as a ReadData object.}
\arguments{
\item{printnames}{List of tests applied to VCF}

\item{results}{Numeric vector of metrics calculated from VCF file}

\item{plots}{List of plots created from VCF File}

\item{tests}{TRUE (passed) or FALSE (failed) logical vector of whether VCF passed metrics using thresholds from VCFQAParam}

\item{fn}{Filename of VCF evaluated (for plot titles)}
}
\description{
Declare class VCFQAReport which will evaluate a VCF stored as a ReadData object.
}
