% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{XHeterozygosity}
\alias{XHeterozygosity}
\title{Calculate X heterozygosity to determine Gender}
\usage{
XHeterozygosity(x)
}
\arguments{
\item{x}{myvcf of interest}
}
\value{
number of X chromosome het calls
}
\description{
Calculate X heterozygosity to determine Gender
}
\keyword{internal}
