% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFQAReport.R
\name{calltypePlot}
\alias{calltypePlot}
\title{Bar plot of variants (counts)}
\usage{
calltypePlot(x, fn)
}
\arguments{
\item{x}{myvcf of interest}

\item{fn}{filename for plot title}
}
\value{
ggplot bar plot of variant types
}
\description{
Bar plot of variants (counts)
}
\keyword{internal}
